# newapp/common/response.go
package common

import (
	"github.com/gin-gonic/gin"
)

// Response defines the professional JSON response format.
type Response struct {
	Code    int         `json:"code"`    // e.g., 200 for success, 400/500 for errors
	Message string      `json:"message"` // descriptive message
	Data    interface{} `json:"data"`    // result payload (if any)
}

// RespondJSON sends a JSON response with the specified code, message, and data.
func RespondJSON(c *gin.Context, code int, message string, data interface{}) {
	// Use the 'code' parameter to set the correct HTTP status code
	c.JSON(code, Response{
		Code:    code,
		Message: message,
		Data:    data,
	})
}